Add files via upload
Checking if parentheses, brackets, and curly brackets are valid by checking if there is a matching one just after it. Right away, we insert a starting bracket into a stack if there is one. If it starts with an ending bracket, immediately break and return false. If there is a starting bracket put into the new stack, it stays, and we redo the process when a corresponding ending bracket is seen, pop out the starting bracket that matches just before it. Once this is done, the stack should be cleared out, so we added if not stack, return True at the end of the function.