Add files via upload
Taking in an int, splitting each digit of that int into a single digit, squaring each digit, and adding them all together. If, eventually, all of those sqaured digits add up to be 1, we break and return True. If not, we return false. Typically, the second condition would be an endless loop, but we are logging each total number and storing them into a list. Each completion runs the current total against all of the items in the list. If the total has already been put into the list, we put a stop to the loop, break, and return false because at that point we understand that it would keep doing the same thing over and over, never hitting 1.