Add files via upload
My first medium problem!!!
Took an integer - copied it to change it around while we go through the code. If the number is 0, return 0 right away. If it is negative, we change to positive right away just so it's easier to work with - we can change at the end. Used mod 10 and divide by 10 to append the last digits of an int into an array. While appending them, changed them to strings so that I could join them. Joined them, then converted them into ints (didn't need to reverse the array btw as you would normally want to because reversing the int is the whole point of the problem). Checked the original int once again - if it's negative, change it back by multiplying by -1. Also checking the edge case mentioned in the description to return 0 if we go beyond a certain int 32 bit int range. Returned the new int.

Fun first medium - can't wait to do more
runtime - 22ms: beats 28.03%
memory - 12.32mb: beats 76.74%