Add files via upload
Took in two strings. Converted them both into arrays. As long as the length of needle + the currently checked index in the array wasn't longer than the entire haystack, went through and verified that the current index and the next ones that would make it the length of the needle itself. If they matched entirely, returned the for loop index that was currently being checked. If they did not and no index+the next few indexes to complete the length of needle matched entirely, returned -1