Add files via upload
We are looking to combine both arrays, sort them and remove non-numbers or fake numbers - whatever you'd like to call them. It seems that only 0's can be fake numbers, so we first append nums2 to nums1, sort that, and then go through this sorted array and remove 0's until c = 0. C refers to the length of each of the arrays minus the amount of actual numbers in each of them (to get the fake amount). Once we remove the fake amount of 0's by popping them out, we have our array ready.
Runtime - 0ms: beats 100%
Memory - 12.35mb: beats 83.36%