Add files via upload
A simple demonstration of what makes python really good. Essentially 3 lines of code to solve a problem. Since count() is already something that exists, we can just use count() == 1 to find the single number within the array. 

If we weren't to use count, we could use nested for loops to compare one index to every other index in the array. If they match, move on to the next number in the array. If there are no matches, break and return that number.