Add files via upload
To be honest, checked the internet a bit before coming up with this solution. My initial idea was to have nested for loops and compare each index with the rest of the array and whenever the second for loop index matched the first, the count would go up and when the count was greater than half the length of the array, we would return the original index. If it never hit, we'd move on and eventually find it given that it is mentioned it always exists at some point. This caused issues with much larger arrays that were used in leetcode test cases, so my idea was abandoned. This answer is much simpler and more of an out of the box approach by sorting and then just returning the middle number as the majority number will always be in the middle after a sort.
Runtime - 3ms: beats 87.11%
Memory - 13.69mb: beats 64%