Add files via upload
Learned a lot about sliding window and time complexity on this one. At first, I tried using one for loop that would add characters that were unique to a list as it went through the loop. Once they weren't unique, I would check if it was the longest, set it to the candidate, and go along through the list, returning the candidate's length at the end. This didn't work, because something like "dvdf" would find the second d, and wipe out the v, which was where the longest substring had to start. To remedy this, I created nested for loops that would start at each index, go through the loop until, again, we saw a duplicate, then move to the next integer. This didn't work because of time constraints on bigger strings. I also wasn't using a set at this point which allows for a faster runtime. I turned to google and learned about sliding window which essentially uses one for loop, goes through the indexes, and includes a start and an end. If the value is a duplicate, move on to the next index and add 1 to the start, remove the duplicate value in the set, and add the new value in along with moving the start. All of these changes allowed much faster processing.

Runtime - 8mb: beats 98.72%
Memory - 13.38mb: beats 17.13%